\docType{data}
\name{seq}
\alias{seq}
\title{
    Example genomic sequence string.
}
\format{A string containing genomic sequence of a ribosomal RNA 18S.}
\usage{
    seq
}
\value{
    Genomic sequence.
}
\description{
    A string containing genomic sequence of a ribosomal RNA 18S.

    This string is provided as a reference for creating a
    \code{SummarizedExperiment} object for storing all structure probing data
    used by the \code{BUMHMM} package. See \code{\link[BUMHMM]{se}} for the
    example code.
}
\references{
    RDN18-1, S000006482, Saccharomyces Genome Database.
}
\keyword{datasets}
