\docType{data}
\name{sequence}
\alias{sequence}
\title{
    Example sequence string.
}
\format{A string containing genetic sequence of a ribosomal RNA 18S.}
\usage{
    sequence
}
\value{
    Genetic sequence.
}
\description{
    A string containing genetic sequence of a ribosomal RNA 18S.
}
\references{
    RDN18-1, S000006482, Saccharomyces Genome Database.
}
\keyword{datasets}
